+++ CHANGELOG.md
@@ -1,16 +1,26 @@
 # Changelog
 
-## [RHa] - Version 1.1.1 - 2019-12-11
+## [RHa] - Version 1.2.1 - 2019-11-14
+### Fix verifyToken() middleware function
+* Add module.exports to auth.js to export verifyToken() to be used by other classes
+
+## [RHa] - Version 1.2.0 - 2019-11-14
+### Add advertisement functionality
+* Add advertisement & comment models
+* Add advertisement routes
+* Add advertisement controller
+
+## [RHa] - Version 1.1.1 - 2019-11-12
 ### Bugfix for endpoint mapping
 * Fix incorrect endpoint mapping
 
-## [RHa] - Version 1.1.0 - 2019-12-11
+## [RHa] - Version 1.1.0 - 2019-11-12
 ### Implement authentication functionality
 * Add user & auth models
 * Add auth routes
 * Add auth controller
 
-## [RHa] - Version 1.0.0 - 2019-12-11
+## [RHa] - Version 1.0.0 - 2019-11-12
 ### Initial setup
 * Define basic project structure
 * Add basic setup & configuration